[CLEANUP] Full Lint Compliance

+++ bifrost/pkg/chainclients/dogecoin/client.go
@@ -1223,7 +1223,7 @@ func (c *Client) ReportSolvency(dogeBlockHeight int64) error {
 			c.logger.Err(err).Msgf("fail to get account balance")
 			continue
 		}
-		if runners.IsVaultSolvent(acct, asgard, cosmos.NewUint(3*EstimateAverageTxSize*uint64(c.lastFeeRate))) && c.IsBlockScannerHealthy() {
+		if runners.IsVaultSolvent(acct, asgard, cosmos.NewUint(3*EstimateAverageTxSize*c.lastFeeRate)) && c.IsBlockScannerHealthy() {
 			// when vault is solvent , don't need to report solvency
 			continue
 		}
